Removing the alias for the farming command, which has become useless due to one of the last updates
