+++ src/main/resources/plugin.yml
@@ -7,6 +7,8 @@ author: SrSyntaxAT
 website: https://srsyntax.at/
 
 commands:
+  farming:
+    description: Teleport you to a farmworld.
   teleportfarmingworld:
     aliases:
       - tpfw
